move log handling to separate file
